improve: change file name for video viewer dialog
